Organize React app
