+++ client/src/components/App/index.js
@@ -1,0 +1,4 @@
+import App from './App';
+import './App.css';
+
+export default App;
